requirements.txt
streamlit>=1.28.0
google-generativeai==0.8.5
PyMuPDF>=1.23.0
python-docx>=0.8.11
pydantic>=2.0.0
fpdf2==2.7.6
Pillow>=10.0.0
python-dateutil>=2.8.2